Add files via upload
This project is a Python-based network sniffer that captures and analyzes network traffic in real-time. Built using the scapy library, it provides insights into how data flows across a network by dissecting and displaying details of IP, TCP, and UDP packets.This project was developed as part of my internship with CodeAlpha, where I gained hands-on experience in network programming and packet analysis using Python. It was a great opportunity to deepen my understanding of networking concepts and apply them in a practical setting.